gate: Use wire entity for data.
