+++ VisualPinball.Unity/VisualPinball.Unity/VPT/Gate/GateBehavior.cs
@@ -4,7 +4,6 @@
 // ReSharper disable MemberCanBePrivate.Global
 #endregion
 
-using Unity.Entities;
 using UnityEngine;
 using VisualPinball.Engine.VPT.Gate;
 using VisualPinball.Unity.Extensions;
@@ -12,29 +11,9 @@ using VisualPinball.Unity.Extensions;
 namespace VisualPinball.Unity.VPT.Gate
 {
 	[AddComponentMenu("Visual Pinball/Gate")]
-	public class GateBehavior : ItemBehavior<Engine.VPT.Gate.Gate, GateData>, IConvertGameObjectToEntity
+	public class GateBehavior : ItemBehavior<Engine.VPT.Gate.Gate, GateData>
 	{
 		protected override string[] Children => new []{"Wire", "Bracket"};
-
-		public void Convert(Entity entity, EntityManager dstManager, GameObjectConversionSystem conversionSystem)
-		{
-			Convert(entity, dstManager);
-
-			dstManager.AddComponentData(entity, new GateStaticData {
-				AngleMin = data.AngleMin,
-				AngleMax = data.AngleMax,
-				Height = data.Height,
-				Damping = data.Damping,
-				GravityFactor = data.GravityFactor,
-				TwoWay = data.TwoWay
-			});
-			dstManager.AddComponentData(entity, new GateMovementData {
-				Angle = data.AngleMin,
-				AngleSpeed = 0,
-				ForcedMove = false,
-				IsOpen = false
-			});
-		}
 
 		protected override Engine.VPT.Gate.Gate GetItem()
 		{
