+++ VisualPinball.Unity/VisualPinball.Unity/VPT/Gate/GateMovementSystem.cs
@@ -36,6 +36,7 @@ namespace VisualPinball.Unity.VPT.Gate
 
 				marker.Begin();
 
+
 				var rotationX = movementData.Angle - math.radians(data.AngleMin);
 				rot.Value = quaternion.RotateX(rotationX);
